+++ Avalonia.Lottie/Animation/Keyframe/GradientColorKeyframeAnimation.cs
@@ -13,10 +13,10 @@ namespace Avalonia.Lottie.Animation.Keyframe
         {
             var startValue = keyframes[0].StartValue;
             var size = startValue?.Size ?? 0;
-            _gradientColor = new GradientColor(new float[size], new Color[size]);
+            _gradientColor = new GradientColor(new double [size], new Color[size]);
         }
 
-        public override GradientColor GetValue(Keyframe<GradientColor> keyframe, float keyframeProgress)
+        public override GradientColor GetValue(Keyframe<GradientColor> keyframe, double  keyframeProgress)
         {
             _gradientColor.Lerp(keyframe.StartValue, keyframe.EndValue, keyframeProgress);
             return _gradientColor;
