+++ Avalonia.Lottie/IInterpolator.cs
@@ -2,6 +2,6 @@ namespace Avalonia.Lottie
 {
     public interface IInterpolator
     {
-        float GetInterpolation(float f);
+        double  GetInterpolation(double f);
     }
 }
